requirements.txt
Sphinx
coverage
flake8
python-dotenv>=0.5.1
pandas
numpy
matplotlib
boto3
dvc
jupyterlab
scikit-learn
argparse
mlflow
django
psycopg2-binary
aws-ecs-services
xgboost